+++ src/entities/discussions/discussions/model/query/discussion.query.type.ts
@@ -1,4 +1,4 @@
-import { TVoteStatus } from '@/entities/vote/model/vote.mutation.type';
+import { TVoteStatus } from '@/entities/discussions/vote/model/vote.mutation.type';
 import { IUserInfo } from '@/shared/types/auth';
 
 /**토론글 get 요청시, 리스폰스(res.data.result) 로 받는 인터페이스  */
