+++ src/entities/discussions/index.ts
@@ -1,13 +1,31 @@
-/**mutation & query */
-export { useDiscussionsQuery } from './model/query/discussions.query';
-export { useCreateDiscussionContent } from './model/mutation/discussions.mutations';
-export { useEditDiscussionContent } from './model/mutation/discussions.mutations';
-export { useDeleteDiscussionContent } from './model/mutation/discussions.mutations';
+/**discussions */
+export { useDiscussionsQuery } from './discussions/model/query/discussions.query';
+export { useCreateDiscussionContent } from './discussions/model/mutation/discussions.mutations';
+export { useEditDiscussionContent } from './discussions/model/mutation/discussions.mutations';
+export { useDeleteDiscussionContent } from './discussions/model/mutation/discussions.mutations';
 
-export { DISCUSSION_CREATE_VALUE } from './model/initial.value';
+export { DISCUSSION_CREATE_VALUE } from './discussions/model/initial.value';
 
-export type { IDiscussionContentMutationRequest } from './model/mutation/discussions.types';
-export type { TDiscussionContentMutationResponse } from './model/mutation/discussions.types';
+export type { IDiscussionContentMutationRequest } from './discussions/model/mutation/discussions.types';
 
-export type { IDiscussionResponse } from './model/query/discussion.query.type';
-export type { IDiscussionContentResponse } from './model/query/discussion.query.type';
+/**reply */
+export { useRepliesQuery } from './reply/model/query/replies.query';
+export { useNestedRepliesQuery } from './reply/model/query/replies.query';
+
+export type { IGetRepliesResponse } from './reply/model/query/replies.query.types';
+export type { IReply } from './reply/model/query/replies.query.types';
+
+export { useCreateReplyMutation } from './reply/model/mutation/replies.mutation';
+export { useEditReplyMutation } from './reply/model/mutation/replies.mutation';
+export { useDeleteReplyMutation } from './reply/model/mutation/replies.mutation';
+
+export type { ICreateReplyMutationRequest } from './reply/model/mutation/replies.mutation.types';
+export type { IEditReplyMutationRequest } from './reply/model/mutation/replies.mutation.types';
+export type { IReplyMutationResponse } from './reply/model/mutation/replies.mutation.types';
+
+/**vote */
+export { useVoteStatusMutation } from './vote/model/vote.mutation';
+export type { IVoteMutationRequest } from './vote/model/vote.mutation.type';
+export type { IVoteMutationResponse } from './vote/model/vote.mutation.type';
+
+export type { TVoteStatus } from './vote/model/vote.mutation.type';
