+++ src/entities/discussions/vote/model/vote.mutation.ts
