+++ src/entities/discussions/vote/model/vote.mutation.type.ts
