--- src/entities/vote/index.ts
@@ -1,5 +1,0 @@
-export { useVoteStatusMutation } from './model/vote.mutation';
-export type { IVoteMutationRequest } from './model/vote.mutation.type';
-export type { IVoteMutationResponse } from './model/vote.mutation.type';
-
-export type { TVoteStatus } from './model/vote.mutation.type';
