+++ src/features/discussions/reply/ui/NestedReply.tsx
@@ -3,8 +3,8 @@ import { ProblemId } from '@/shared';
 import { BouncingDots } from '@/shared/ui/loading-indicators';
 import { useState } from 'react';
 import Vote from '../../vote/ui/Vote';
-import { IReply, useDeleteReplyMutation } from '@/entities/discussionReplies';
 import ReplyForm from './ReplyForm';
+import { IReply, useDeleteReplyMutation } from '@/entities/discussions';
 
 interface INestedReplyProps {
   nestedReply: IReply;
