+++ src/features/discussions/reply/ui/Reply.tsx
@@ -4,9 +4,9 @@ import { ProblemId } from '@/shared';
 import { useState } from 'react';
 import { BouncingDots } from '@/shared/ui/loading-indicators';
 import NestedReplies from './NestedReplies';
-import { IReply, useDeleteReplyMutation } from '@/entities/discussionReplies';
 import ReplyForm from './ReplyForm';
 import ShowChildReplies from './ShowChildReplies';
+import { IReply, useDeleteReplyMutation } from '@/entities/discussions';
 
 interface IReplyProps {
   reply: IReply;
